requirements.txt
streamlit
pandas
streamlit-aggrid
psycopg2-binary
sqlalchemy
